Show preview when sharing media from intent
