+++ features/mediapreview/impl/build.gradle.kts
@@ -2,8 +2,7 @@ import extension.setupDependencyInjection
 import extension.testCommonDependencies
 
 /*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2023, 2024 New Vector Ltd.
+ * Copyright (c) 2026 Element Creations Ltd.
  *
  * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
  * Please see LICENSE files in the repository root for full details.
@@ -15,8 +14,7 @@ plugins {
 }
 
 android {
-    namespace = "io.element.android.features.share.impl"
-
+    namespace = "io.element.android.features.mediapreview.impl"
     testOptions {
         unitTests {
             isIncludeAndroidResources = true
@@ -27,32 +25,33 @@ android {
 setupDependencyInjection()
 
 dependencies {
+    api(projects.features.mediapreview.api)
     implementation(projects.appconfig)
     implementation(projects.libraries.androidutils)
     implementation(projects.libraries.core)
     implementation(projects.libraries.architecture)
-    implementation(projects.libraries.featureflag.api)
     implementation(projects.libraries.matrix.api)
     implementation(projects.libraries.matrixui)
     implementation(projects.libraries.designsystem)
     implementation(projects.libraries.mediaupload.api)
-    implementation(projects.libraries.roomselect.api)
-    implementation(projects.libraries.uiStrings)
-    implementation(projects.libraries.testtags)
-    implementation(projects.services.appnavstate.api)
+    implementation(projects.libraries.mediaupload.impl)
     implementation(projects.libraries.mediaviewer.api)
     implementation(projects.libraries.textcomposer.impl)
-    implementation(projects.features.mediapreview.api)
-    implementation(projects.features.mediapreview.impl)
+    implementation(projects.libraries.uiStrings)
+    implementation(projects.libraries.preferences.api)
+    implementation(projects.libraries.di)
+    implementation(projects.libraries.uiUtils)
+    implementation(projects.libraries.featureflag.api)
+    implementation(libs.androidx.compose.foundation)
+    implementation(libs.androidx.compose.foundation.layout)
+    implementation(libs.androidx.exifinterface)
+    implementation(libs.coil.compose)
     implementation(libs.timber)
-    api(projects.features.share.api)
-    api(libs.statemachine)
 
     testCommonDependencies(libs, true)
-    testImplementation(projects.features.share.test)
     testImplementation(projects.libraries.matrix.test)
     testImplementation(projects.libraries.mediaupload.test)
     testImplementation(projects.libraries.preferences.test)
-    testImplementation(projects.libraries.roomselect.test)
-    testImplementation(projects.services.appnavstate.impl)
+    testImplementation(projects.libraries.mediaviewer.test)
+    testImplementation(projects.libraries.featureflag.test)
 }
