+++ features/mediapreview/impl/src/main/kotlin/io/element/android/features/mediapreview/impl/MediaPreviewEvents.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2026 Element Creations Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package io.element.android.features.mediapreview.impl
+
+import io.element.android.features.mediapreview.impl.imageeditor.NormalizedCropRect
+import io.element.android.libraries.preferences.api.store.VideoCompressionPreset
+
+sealed interface MediaPreviewEvents {
+    data object Send : MediaPreviewEvents
+    data object Cancel : MediaPreviewEvents
+    data object Retry : MediaPreviewEvents
+    data object ClearError : MediaPreviewEvents
+    data class ToggleImageOptimization(val enabled: Boolean) : MediaPreviewEvents
+    data class SelectVideoQuality(val preset: VideoCompressionPreset) : MediaPreviewEvents
+    data object ShowVideoQualityDialog : MediaPreviewEvents
+    data object DismissVideoQualityDialog : MediaPreviewEvents
+    data object OpenImageEditor : MediaPreviewEvents
+    data object CloseImageEditor : MediaPreviewEvents
+    data object RotateImageToTheLeft : MediaPreviewEvents
+    data object ApplyImageEdits : MediaPreviewEvents
+    data object ResetImageEdits : MediaPreviewEvents
+    data class UpdateImageCropRect(val cropRect: NormalizedCropRect) : MediaPreviewEvents
+    data object ClearImageEditError : MediaPreviewEvents
+}
