--- features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/attachments/preview/AttachmentsPreviewNode.kt
@@ -1,79 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2023-2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.features.messages.impl.attachments.preview
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import com.bumble.appyx.core.modality.BuildContext
-import com.bumble.appyx.core.node.Node
-import com.bumble.appyx.core.plugin.Plugin
-import dev.zacsweers.metro.Assisted
-import dev.zacsweers.metro.AssistedInject
-import io.element.android.annotations.ContributesNode
-import io.element.android.compound.colors.SemanticColorsLightDark
-import io.element.android.compound.theme.ForcedDarkElementTheme
-import io.element.android.features.enterprise.api.EnterpriseService
-import io.element.android.features.messages.impl.attachments.Attachment
-import io.element.android.libraries.architecture.NodeInputs
-import io.element.android.libraries.architecture.inputs
-import io.element.android.libraries.di.RoomScope
-import io.element.android.libraries.matrix.api.core.EventId
-import io.element.android.libraries.matrix.api.core.SessionId
-import io.element.android.libraries.matrix.api.timeline.Timeline
-import io.element.android.libraries.mediaviewer.api.local.LocalMediaRenderer
-
-@ContributesNode(RoomScope::class)
-@AssistedInject
-class AttachmentsPreviewNode(
-    @Assisted buildContext: BuildContext,
-    @Assisted plugins: List<Plugin>,
-    presenterFactory: AttachmentsPreviewPresenter.Factory,
-    private val localMediaRenderer: LocalMediaRenderer,
-    private val sessionId: SessionId,
-    private val enterpriseService: EnterpriseService,
-) : Node(buildContext, plugins = plugins) {
-    data class Inputs(
-        val attachment: Attachment,
-        val timelineMode: Timeline.Mode,
-        val inReplyToEventId: EventId?,
-    ) : NodeInputs
-
-    private val inputs: Inputs = inputs()
-
-    private val onDoneListener = OnDoneListener {
-        navigateUp()
-    }
-
-    private val presenter = presenterFactory.create(
-        attachment = inputs.attachment,
-        timelineMode = inputs.timelineMode,
-        onDoneListener = onDoneListener,
-        inReplyToEventId = inputs.inReplyToEventId,
-    )
-
-    @Composable
-    override fun View(modifier: Modifier) {
-        val colors by remember {
-            enterpriseService.semanticColorsFlow(sessionId = sessionId)
-        }.collectAsState(SemanticColorsLightDark.default)
-        ForcedDarkElementTheme(
-            colors = colors,
-        ) {
-            val state = presenter.present()
-            AttachmentsPreviewView(
-                state = state,
-                localMediaRenderer = localMediaRenderer,
-                modifier = modifier
-            )
-        }
-    }
-}
