+++ libraries/mediaupload/test/src/main/kotlin/io/element/android/libraries/mediaupload/test/FakeVideoMetadataExtractor.kt
@@ -1,12 +1,11 @@
 /*
  * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2025 New Vector Ltd.
  *
  * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.features.messages.test.attachments.video
+package io.element.android.libraries.mediaupload.test
 
 import android.net.Uri
 import android.util.Size
@@ -15,8 +14,8 @@ import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 
 class FakeVideoMetadataExtractor(
-    private val sizeResult: Result<Size> = Result.success(Size(1, 1)),
-    private val duration: Result<Duration> = Result.success(1.milliseconds),
+    private val sizeResult: Result<Size> = Result.success(Size(1920, 1080)),
+    private val duration: Result<Duration> = Result.success(10_000.milliseconds),
 ) : VideoMetadataExtractor {
     override fun getSize(): Result<Size> = sizeResult
