+++ tests/uitests/src/test/snapshots/images/features.mediapreview.impl_MediaPreviewView_Day_0_en.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:33db404c55957762e348c1519c37f8c7664576d573403607d41928e601c75a8f
+size 13849
