+++ tests/uitests/src/test/snapshots/images/features.mediapreview.impl_MediaPreviewView_Night_0_en.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:666c7914865a2641fdc5d8f318e0d23282180a5eb4480cd72688eb1c672c2e5c
+size 12940
